requirements.txt
streamlit==1.38.0
pandas==2.2.3
python-dotenv==1.0.1
newsapi-python==0.2.7
flask==3.0.3
pywebpush==2.0.0
gunicorn==23.0.0